Next up: finish removal of explicit FCOTMR manipulation from TypedTokenStore
