+++ hedera-node/src/main/java/com/hedera/services/ServicesState.java
@@ -67,7 +67,6 @@ import java.util.function.Supplier;
 
 import static com.hedera.services.context.SingletonContextsManager.CONTEXTS;
 import static com.hedera.services.sigs.HederaToPlatformSigOps.expandIn;
-import static com.hedera.services.state.initialization.ViewBuilder.rebuildUniqueTokenViews;
 import static com.hedera.services.state.merkle.MerkleNetworkContext.UNKNOWN_CONSENSUS_TIME;
 import static com.hedera.services.utils.EntityIdUtils.asLiteralString;
 import static com.hedera.services.utils.EntityIdUtils.parseAccount;
@@ -281,7 +280,7 @@ public class ServicesState extends AbstractNaryMerkleInternal implements SwirldS
 		uniqueTokenAssociations = new FCOneToManyRelation<>();
 		uniqueOwnershipAssociations = new FCOneToManyRelation<>();
 		uniqueTreasuryOwnedAssociations = new FCOneToManyRelation<>();
-		rebuildUniqueTokenViews(ctx.tokenStore(), uniqueTokens(), uniqueTokenAssociations, uniqueOwnershipAssociations, uniqueTreasuryOwnedAssociations);
+		ctx.uniqTokenViewsManager().rebuildNotice(tokens(), uniqueTokens());
 		/* Use any payer records stored in state to rebuild the recent transaction
 		 * history. This history has two main uses: Purging expired records, and
 		 * classifying duplicate transactions. */
