+++ hedera-node/src/test/java/com/hedera/services/ServicesStateTest.java
@@ -33,7 +33,6 @@ import com.hedera.services.records.TxnIdRecentHistory;
 import com.hedera.services.sigs.order.HederaSigningOrder;
 import com.hedera.services.sigs.order.SigningOrderResult;
 import com.hedera.services.state.expiry.ExpiryManager;
-import com.hedera.services.state.initialization.ViewBuilderTest;
 import com.hedera.services.state.logic.NetworkCtxManager;
 import com.hedera.services.state.merkle.MerkleAccount;
 import com.hedera.services.state.merkle.MerkleBlobMeta;
@@ -52,6 +51,7 @@ import com.hedera.services.state.submerkle.EntityId;
 import com.hedera.services.state.submerkle.ExchangeRates;
 import com.hedera.services.state.submerkle.SequenceNumber;
 import com.hedera.services.store.tokens.TokenStore;
+import com.hedera.services.store.tokens.views.UniqTokenViewsManager;
 import com.hedera.services.stream.RecordStreamManager;
 import com.hedera.services.stream.RecordsRunningHashLeaf;
 import com.hedera.services.throttling.FunctionalityThrottling;
@@ -521,10 +521,11 @@ class ServicesStateTest {
 	@Test
 	void rebuildsFcotmrAsExpected() {
 		// setup:
+		final var uniqTokenViewsManager = mock(UniqTokenViewsManager.class);
 		var nodeInfo = mock(NodeInfo.class);
 		given(ctx.nodeInfo()).willReturn(nodeInfo);
 		TokenStore tokenStore = mock(TokenStore.class);
-		given(ctx.tokenStore()).willReturn(tokenStore);
+		given(ctx.uniqTokenViewsManager()).willReturn(uniqTokenViewsManager);
 		given(nodeInfo.selfAccount()).willReturn(nodeAccount);
 		CONTEXTS.store(ctx);
 
@@ -539,14 +540,12 @@ class ServicesStateTest {
 		subject.setChild(ServicesState.ChildIndices.DISK_FS, diskFs);
 		subject.setChild(ServicesState.ChildIndices.SCHEDULE_TXS, scheduledTxs);
 		subject.setChild(ServicesState.ChildIndices.RECORD_STREAM_RUNNING_HASH, runningHashLeaf);
-		subject.setChild(ServicesState.ChildIndices.UNIQUE_TOKENS, ViewBuilderTest.someUniqueTokens());
+		subject.setChild(ServicesState.ChildIndices.UNIQUE_TOKENS, uniqueTokens);
 		// when:
 		subject.init(platform, book);
 
 		// then:
-		ViewBuilderTest.assertIsTheExpectedUta(subject.uniqueTokenAssociations());
-		ViewBuilderTest.assertIsTheExpectedUtao(subject.uniqueOwnershipAssociations());
-//		ViewBuilderTest.assertIsTheExpectedTUtao(subject.uniqueOwnershipTreasuryAssociations());  TODO case with ownership treasury
+		verify(uniqTokenViewsManager).rebuildNotice(tokens, uniqueTokens);
 	}
 
 	@Test
