+++ hedera-node/src/test/java/com/hedera/services/store/tokens/views/utils/MultiSourceRangeTest.java
@@ -1,8 +1,28 @@
-package com.hedera.services.store.tokens.utils;
+package com.hedera.services.store.tokens.views.utils;
+
+/*-
+ * ‌
+ * Hedera Services Node
+ * ​
+ * Copyright (C) 2018 - 2021 Hedera Hashgraph, LLC
+ * ​
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ‍
+ */
 
 import org.junit.jupiter.api.Test;
 
-import static com.hedera.services.store.tokens.utils.MultiSourceRange.EMPTY_RANGE;
+import static com.hedera.services.store.tokens.views.utils.MultiSourceRange.EMPTY_RANGE;
 import static org.junit.jupiter.api.Assertions.*;
 
 class MultiSourceRangeTest {
